feat: implement Forward-Forward and RTRL — biologically plausible alternatives to backpropagation
Add two biologically-inspired learning algorithms as alternatives to BPTT:

Forward-Forward (Hinton 2022):
- src/forward_forward.rs: L2 normalisation, goodness, FF loss functions
- src/ff_model.rs: FfLayer, FfEncoder with stackable per-layer local learning
- circle-world-demo/src/ff_main.rs: FF autoencoder demo

RTRL — Real-Time Recurrent Learning (zero BPTT):
- src/rtrl.rs: ssm_step_detached, RtrlAccumulator — 1-step truncated RTRL
- src/rtrl_world.rs: RtrlWorldModel (FF enc + RTRL SSM + FF dec)
- circle-world-demo/src/rtrl_main.rs: RTRL demo — proven viable

Documentation:
- docs/BEYOND_BACKPROP.md: comprehensive research doc covering motivation,
  architecture, experiments, failure modes, and future directions

Refactor:
- src/lib.rs: register new modules
- src/ssm.rs: minor improvements
- nab-demo/src/main.rs: refactor for clarity
- Cargo.toml: add circle-world-rtrl binary target

Key finding: RTRL v1 works — the SSM learns to predict latent dynamics and
FF encoder/decoder learn to represent the circle, all with zero BPTT and
zero cross-component BP. This is 'Oil' on the path to 'Nuclear Power'
(Predictive Coding / Equilibrium Propagation).

+++ src/ssm.rs
@@ -116,7 +116,7 @@ impl<B: Backend> MultiScaleSsmBlock<B> {
             // Override a_re with layer-specific initialization
             block.a_re = Param::from_tensor(Tensor::random(
                 [config.n_heads, config.d_state],
-                Distribution::Uniform(a_re_min as f64, a_re_max as f64),
+                Distribution::Uniform(a_re_min, a_re_max),
                 device,
             ));
 
@@ -147,8 +147,7 @@ impl<B: Backend> MultiScaleSsmBlock<B> {
             let normalized = norm.forward(out);
             let y = layer.forward(normalized);
             // Scale residual contribution to prevent activation explosion
-            out = (y.mul_scalar(num_layers_recip) + residual)
-                .clamp(-100.0, 100.0);
+            out = (y.mul_scalar(num_layers_recip) + residual).clamp(-100.0, 100.0);
         }
         out
     }
@@ -177,8 +176,7 @@ impl<B: Backend> MultiScaleSsmBlock<B> {
                     None
                 },
             );
-            out = (y.mul_scalar(num_layers_recip) + residual)
-                .clamp(-100.0, 100.0);
+            out = (y.mul_scalar(num_layers_recip) + residual).clamp(-100.0, 100.0);
             new_states.push((next_h, current_bx));
             new_conv_states.push(next_conv_state);
         }
@@ -208,6 +206,7 @@ pub struct MultiScaleState<B: Backend> {
 }
 
 impl<B: Backend> MultiScaleState<B> {
+    #[allow(clippy::too_many_arguments)]
     pub fn zeros(
         batch: usize,
         n_layers: usize,
